move sort functionality to the contact page object
